[QA-491] Possible problems implementation (#4)
* Implement test cases for different problems, modified program and test structure accordingly. Added documentation

* Small fix

* Program optimization, added header transfer for pipe

* Header transfer optimization for pipe, get option optimization, getFileSize() return ssize_t

* Minor fix

* A beta version for Ipc_method is added to handle file size bigger than 2GB for pipe, Ipc_pipe optimazation

* Fixed buffer size issue with pipe

+++ Cpp_hands_on_programming/testReceive/testReceive.cc
@@ -7,6 +7,8 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 #include <fstream>
+#include "testlib/PolymorphismInput.h"
+#include "testlib/CaptureStdOut.h"
 
 std::map<std::string, std::shared_ptr<Ipc_method>> input_ipc_pointer = {
     {"pipe", std::make_shared<Ipc_pipe>("gpipe", "pipeGtest")},
@@ -14,10 +16,12 @@ std::map<std::string, std::shared_ptr<Ipc_method>> input_ipc_pointer = {
     {"queue", std::make_shared<Ipc_queue>("/gqueue", "queueGtest")},
     {"queueBigfile", std::make_shared<Ipc_queue>("/gqueue2", "queueBigfileGtest")},
     {"shm", std::make_shared<Ipc_shm>("/gshm", "shmGtest")},
-    {"shmBigfile", std::make_shared<Ipc_shm>("/gshm2", "shmBigfileGtest")}
-};
+    {"shmBigfile", std::make_shared<Ipc_shm>("/gshm2", "shmBigfileGtest")},
+    {"timeoutPipe", std::make_shared<Ipc_pipe>("ptimeoutReceive", "file1", 1000)},
+    {"timeoutQueue", std::make_shared<Ipc_queue>("/qtimeoutReceive", "file2", 1000)},
+    {"timeoutShm", std::make_shared<Ipc_shm>("/stimeoutReceive", "file3", 1000)}
 
-class PolymorphismInput : public::testing::TestWithParam<std::pair<const std::string, std::shared_ptr<Ipc_method>>> {};
+};
 
 INSTANTIATE_TEST_SUITE_P(MultipleIPCMethods, PolymorphismInput, 
 testing::ValuesIn(input_ipc_pointer), 
@@ -26,30 +30,31 @@ testing::ValuesIn(input_ipc_pointer),
 });
 
 TEST_P(PolymorphismInput, ReceiveData) {
-    const int bigSize = 1000*1000*1000;
-    size_t pos;
-    std::string expected = "Write to file from ";
-    expected += GetParam().first;
-    if ((pos = expected.find("Bigfile")) != std::string::npos) {
-        expected = expected.substr(0, pos); 
-    } 
-    expected +=  " successfully, exiting the program..\n";
-
-    std::stringstream strm;
-    auto cout_stream_buf = std::cout.rdbuf(); // save cout stream buffer
-    // redirect cout to string stream strm
-    std::cout.rdbuf(strm.rdbuf());
-    GetParam().second->receive();
-
-    // reset cout stream buffer to original
-    std::cout.rdbuf(cout_stream_buf);
-
-    std::string file = GetParam().first + "Gtest";
-    if (GetParam().first.find("Bigfile") != std::string::npos) {
-        EXPECT_THAT(GetParam().second->getFileSize(file), testing::Eq(bigSize)); //// size of bigfile file (sent by PipeSendTest)
+    std::string coutString;
+    if (GetParam().first.find("timeout") == std::string::npos) {
+        size_t pos;
+        std::string expected = "Write to file from ";
+        expected += GetParam().first;
+        if ((pos = expected.find("Bigfile")) != std::string::npos) {
+            expected = expected.substr(0, pos); 
+        } 
+        expected +=  " successfully, exiting the program..\n";
+        {
+            CaptureStdOut capture;
+            GetParam().second->receive();
+            coutString = capture.getString();
+        }
+            std::string file = GetParam().first + "Gtest";
+            if (GetParam().first.find("Bigfile") != std::string::npos) {
+            EXPECT_THAT(GetParam().second->getFileSize(file), testing::Eq(bigSize));
+        } else {
+            EXPECT_THAT(GetParam().second->getFileSize(file), testing::Eq(smallSize)); 
+        }
+        EXPECT_THAT(coutString, testing::EndsWith(expected)); 
+        unlink(file.c_str());
+        
     } else {
-        EXPECT_THAT(GetParam().second->getFileSize(file), testing::Eq(174536)); //174536 is the file size of test2 (sent by PipeSendTest)
+        CaptureStdOut capture;
+        EXPECT_THROW(GetParam().second->receive(), std::runtime_error);
     }
-    EXPECT_THAT(strm.str(), testing::EndsWith(expected)); 
-    unlink(file.c_str());
 }
